+++ REvoDesign/tools/system_tools.py
@@ -1,4 +1,3 @@
-import os
 from absl import logging
 import importlib